Extract Folder Submenu to dedicated module
Extracted the folder submenu logic (creation, positioning, toggling) from `prompt-list.js` to a new module `folder-submenu.js`.
This improves code organization and separation of concerns.

- Created `src/modules/folder-submenu.js`
- Refactored `src/modules/prompt-list.js` to use the new module
- Verified with frontend script using mock data.
